+++ pkg/database/delete_test.go
